Adjusted to mention only entry validation and the new syntax for the validation package as was build in the scope of https://github.com/holochain/holochain-rust/pull/503 and https://github.com/holochain/holochain-rust/pull/512.
