+++ zomes/blog/code/Cargo.toml
@@ -8,7 +8,6 @@ serde = "1.0"
 serde_json = "1.0"
 hdk = { git = "https://github.com/holochain/holochain-rust" , branch = "develop" }
 serde_derive = "1.0"
-holochain_wasm_utils = { git = "https://github.com/holochain/holochain-rust" , branch = "develop" }
 boolinator = "2.4"
 
 [lib]
